efi_runtime: Refactor ioctl code into helper functions
efi_runtime_ioctl() has grown to be fairly unwieldy because it includes
all the data objects and logic in one place. Make use of helper
functions for each of the ioctl commands to make the code easier to read
and maintain.

There is no intended change to functionality, just code movement.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>