efi_runtime: Group kernel version dependent functions
Instead of using multiple identical #if preprocessor blocks group the
function calls and implementations together so that we can keep the
preprocessor checks to a minimum.

This change also more clearly conveys the message that the two functions
are dependent on the same kernel version. With multiple #if block it's
easy to misread them and think that, becuase there's more than one, they
must have different kernel version requirements.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>