efi_runtime: do not dereference userspace pointers from strings
__ucs2_strsize is only used by copy_ucs2_from_user to determine the
length of a string coming from the user space. Thus, dereferencing
any part of the string in search of '\0' clearly constitutes an
illegal read of userspace memory from the kernel. Thus, ready each
character of the string using get_user.

Also, while there, add the attribute __user to the argument of the
function for the reason explained above.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>