efi_runtime: allow large sizes for variable names
The UEFI specification does not define the maximum length for the
variable name. Thus, it may happen that some firmware implementations
have variable names longer than 1024 characters. Rather than limiting
the maximum size to 1024 characters, set the initial size to 1024 chars
and enlarge as required.

To allow for longer names, allocate and and grow the allocated memory
as required by the firmware. If we are unable to allocate the needed
memory, we skip the test.

This functionality required to rework the error paths of the involved
functions.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>